requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.4
django-crispy-forms==1.14.0
django-registration-redux==2.11
Pillow==9.4.0
sqlparse==0.4.3
